# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hassette Web API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/logs/recent": {
      "get": {
        "tags": [
          "logs"
        ],
        "summary": "Get Logs",
        "description": "Return recent log records from the database with optional filtering.",
        "operationId": "get_logs_api_logs_recent_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 2000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "app_key",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "App Key"
            }
          },
          {
            "name": "level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Level"
            }
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          },
          {
            "name": "execution_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Execution Id"
            }
          },
          {
            "name": "source_tier",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by source tier: 'app' or 'framework'. Omit to return records from both. Unlike the telemetry-metrics endpoints, which default to 'app', this default deliberately includes everything: the log viewer is a raw feed, not an app-author-facing metric.",
              "title": "Source Tier"
            },
            "description": "Filter by source tier: 'app' or 'framework'. Omit to return records from both. Unlike the telemetry-metrics endpoints, which default to 'app', this default deliberately includes everything: the log viewer is a raw feed, not an app-author-facing metric."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LogEntryResponse"
                  },
                  "title": "Response Get Logs Api Logs Recent Get"
                }
              }
            }
          },
          "503": {
            "description": "`telemetry_unavailable`: the telemetry store could not be read",
            "x-problem-codes": [
              "telemetry_unavailable"
            ],
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/logs/level": {
      "put": {
        "tags": [
          "logs"
        ],
        "summary": "Set Log Level",
        "description": "Change a logger's effective level at runtime without restarting.\n\nThe change takes effect immediately for both structlog and stdlib callers on that logger.",
        "operationId": "set_log_level_api_logs_level_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogLevelRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogLevelResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/scheduler/jobs": {
      "get": {
        "tags": [
          "scheduler"
        ],
        "summary": "All Jobs",
        "description": "All scheduled jobs across all apps, enriched with live registry data.\n\n``schedule_status``/``schedule_status_reason`` and, for ``SCHEDULED`` jobs, live timing\n(``next_run``, ``fire_at``, ``jitter``) are joined from the live scheduler registry by\n``db_id``. On registry failure the DB rows are returned without enrichment (degraded but\nfunctional; logged warning, no 500).\n\nThe registry snapshot is taken once \u2014 not per app \u2014 to avoid fan-out overhead.",
        "operationId": "all_jobs_api_scheduler_jobs_get",
        "parameters": [
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          },
          {
            "name": "source_tier",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "app",
                "framework",
                "all"
              ],
              "type": "string",
              "description": "Filter by source tier. 'app' excludes framework internals. 'framework' returns only internal actors. 'all' returns everything.",
              "default": "app",
              "title": "Source Tier"
            },
            "description": "Filter by source tier. 'app' excludes framework internals. 'framework' returns only internal actors. 'all' returns everything."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobSummary"
                  },
                  "title": "Response All Jobs Api Scheduler Jobs Get"
                }
              }
            }
          },
          "503": {
            "description": "`telemetry_unavailable`: the telemetry store could not be read",
            "x-problem-codes": [
              "telemetry_unavailable"
            ],
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/scheduler/jobs/{job_id}/trigger": {
      "post": {
        "tags": [
          "scheduler"
        ],
        "summary": "Trigger Job",
        "description": "Manually submit a job for immediate execution.\n\nResolves the job by ``job_id`` (the job's ``db_id``) in the live scheduler registry, then\nsubmits it through the same ``SchedulerService.submit_job()`` path used by ``Job.submit()``\n\u2014 fire-and-observe, dispatched via ``run_job_with_guard(job, trigger_mode=\"manual\")``.\n\nA live registration always returns 202 accepted, even when overlap policy (``single``\nmode, queue capacity) later suppresses or drops the invocation \u2014 those outcomes are\ndecided asynchronously by the job's existing guard and are not previewed here. Returns\n409 when the job has no live registration: never registered, or removed via\n``Job.remove()``, ``Scheduler.remove_job()``/``remove_group()``, owner shutdown, or\n``if_exists=\"replace\"`` since the caller last saw it.\n\nManual submission never consumes, moves, or completes a pending automatic occurrence \u2014\na pending one-shot or recurring schedule still fires at its own time.",
        "operationId": "trigger_job_api_scheduler_jobs__job_id__trigger_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobTriggerResponse"
                }
              }
            }
          },
          "409": {
            "description": "`job_not_registered`: the job has no live registration",
            "x-problem-codes": [
              "job_not_registered"
            ],
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ExecutionMode": {
        "type": "string",
        "enum": [
          "single",
          "restart",
          "queued",
          "parallel"
        ],
        "title": "ExecutionMode",
        "description": "Overlap behavior for a listener when a trigger fires while a prior invocation still runs."
      },
...
      "JobSummary": {
        "properties": {
          "job_id": {
            "type": "integer",
            "title": "Job Id"
          },
          "app_key": {
            "type": "string",
            "title": "App Key"
          },
          "instance_index": {
            "type": "integer",
            "title": "Instance Index"
          },
          "job_name": {
            "type": "string",
            "title": "Job Name"
          },
          "handler_method": {
            "type": "string",
            "title": "Handler Method"
          },
          "trigger_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trigger Type"
          },
          "trigger_label": {
            "type": "string",
            "title": "Trigger Label",
            "default": ""
          },
          "trigger_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trigger Detail"
          },
          "args_json": {
            "type": "string",
            "title": "Args Json"
          },
          "kwargs_json": {
            "type": "string",
            "title": "Kwargs Json"
          },
          "source_location": {
            "type": "string",
            "title": "Source Location"
          },
          "registration_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Registration Source"
          },
          "source_tier": {
            "type": "string",
            "enum": [
              "app",
              "framework"
            ],
            "title": "Source Tier",
            "default": "app"
          },
          "predicate_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Predicate Description"
          },
          "human_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Human Description"
          },
          "total_executions": {
            "type": "integer",
            "title": "Total Executions"
          },
          "successful": {
            "type": "integer",
            "title": "Successful"
          },
          "failed": {
            "type": "integer",
            "title": "Failed"
          },
          "cancelled": {
            "type": "integer",
            "title": "Cancelled",
            "default": 0
          },
          "timed_out": {
            "type": "integer",
            "title": "Timed Out",
            "default": 0
          },
          "skipped": {
            "type": "integer",
            "title": "Skipped",
            "default": 0
          },
          "thread_leaked": {
            "type": "integer",
            "title": "Thread Leaked",
            "default": 0
          },
          "last_executed_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Executed At"
          },
          "total_duration_ms": {
            "type": "number",
            "title": "Total Duration Ms"
          },
          "avg_duration_ms": {
            "type": "number",
            "title": "Avg Duration Ms"
          },
          "group": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group"
          },
          "schedule_status": {
            "type": "string",
            "title": "Schedule Status",
            "default": "scheduled"
          },
          "schedule_status_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schedule Status Reason"
          },
          "next_run": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Run"
          },
          "fire_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fire At"
          },
          "jitter": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jitter"
          },
          "last_error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error Message"
          },
          "last_error_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error Type"
          },
          "last_error_ts": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error Ts"
          },
          "last_error_traceback": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error Traceback"
          },
          "min_duration_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Duration Ms"
          },
          "max_duration_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Duration Ms"
          },
          "mode": {
            "$ref": "#/components/schemas/ExecutionMode",
            "default": "single"
          },
          "suppressed_count": {
            "type": "integer",
            "title": "Suppressed Count",
            "default": 0
          },
          "dropped_count": {
            "type": "integer",
            "title": "Dropped Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "app_key",
          "instance_index",
          "job_name",
          "handler_method",
          "trigger_type",
          "args_json",
          "kwargs_json",
          "source_location",
          "registration_source",
          "total_executions",
          "successful",
          "failed",
          "last_executed_at",
          "total_duration_ms",
          "avg_duration_ms"
        ],
        "title": "JobSummary",
        "description": "Per-job summary returned by ``get_job_summary()``.\n\n``failed`` counts only ``'error'`` status; ``timed_out``, ``cancelled``, and ``skipped``\nare tracked separately.\nInvariant: ``successful + failed + cancelled + timed_out + skipped == total_executions``."
      },
      "JobTriggerResponse": {
        "properties": {
          "status": {
            "type": "string",
            "const": "accepted",
            "title": "Status",
            "default": "accepted"
          },
          "job_id": {
            "type": "integer",
            "title": "Job Id"
          },
          "job_name": {
            "type": "string",
            "title": "Job Name"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "job_name"
        ],
        "title": "JobTriggerResponse",
        "description": "Response for POST /api/scheduler/jobs/{job_id}/trigger \u2014 manual job submission.\n\nSeparate from ``ActionResponse`` (which has ``app_key``/``action`` but no ``job_id``) \u2014\nthe trigger response identifies the job, not an app action. ``status`` is always\n``\"accepted\"`` for a live registration (the endpoint returns 409 instead of this model\nwhen the job has no live registration) \u2014 overlap-policy suppression or dropping happens\nasynchronously and is not previewed in this response."
      },
...
      "LogEntryResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "seq": {
            "type": "integer",
            "title": "Seq"
          },
          "timestamp": {
            "type": "number",
            "title": "Timestamp"
          },
          "level": {
            "type": "string",
            "enum": [
              "DEBUG",
              "INFO",
              "WARNING",
              "ERROR",
              "CRITICAL"
            ],
            "title": "Level"
          },
          "logger_name": {
            "type": "string",
            "title": "Logger Name"
          },
          "func_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Func Name"
          },
          "lineno": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lineno"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "exc_info": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exc Info"
          },
          "app_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "App Key"
          },
          "execution_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Id"
          },
          "instance_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance Name"
          },
          "instance_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance Index"
          },
          "source_tier": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "app",
                  "framework"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Tier"
          },
          "execution_kind": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "handler",
                  "job"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Kind"
          },
          "listener_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Listener Id"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "seq",
          "timestamp",
          "level",
          "logger_name",
          "message"
        ],
        "title": "LogEntryResponse"
      },
      "LogLevelRequest": {
        "properties": {
          "logger": {
            "type": "string",
            "title": "Logger"
          },
          "level": {
            "type": "string",
            "title": "Level"
          }
        },
        "type": "object",
        "required": [
          "logger",
          "level"
        ],
        "title": "LogLevelRequest",
        "description": "Request body for PUT /api/logs/level."
      },
      "LogLevelResponse": {
        "properties": {
          "logger": {
            "type": "string",
            "title": "Logger"
          },
          "effective_level": {
            "type": "string",
            "title": "Effective Level"
          }
        },
        "type": "object",
        "required": [
          "logger",
          "effective_level"
        ],
        "title": "LogLevelResponse",
        "description": "Response for PUT /api/logs/level."
      },
...
      "ProblemCode": {
        "type": "string",
        "enum": [
          "invalid_app_key",
          "app_not_found",
          "instance_not_found",
          "bootstrap_not_released",
          "app_blocked",
          "action_failed",
          "telemetry_unavailable",
          "source_not_found",
          "path_traversal",
          "source_unavailable",
          "invalid_token",
          "not_authenticated",
          "job_not_registered",
          "validation_failed",
          "body_too_large",
          "not_found",
          "method_not_allowed",
          "http_error",
          "internal_error"
        ],
        "title": "ProblemCode",
        "description": "Machine-readable reason carried in the ``code`` member of every web API error body.\n\nEach code maps to exactly one HTTP status, except ``http_error``, which carries whatever\nstatus the underlying error had. Before 1.0, renaming or removing a code, or changing its\nstatus, is a breaking change and is flagged as one; adding a code is not."
      },
      "ProblemDetail": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Problem type URI. Always ``about:blank``: ``code`` carries the specific meaning.",
            "default": "about:blank"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "The HTTP reason phrase for ``status``."
          },
          "status": {
            "type": "integer",
            "title": "Status",
            "description": "The HTTP status code of the response."
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Human-readable explanation of this occurrence."
          },
          "code": {
            "$ref": "#/components/schemas/ProblemCode",
            "description": "Machine-readable reason for the error."
          }
        },
        "type": "object",
        "required": [
          "title",
          "status",
          "detail",
          "code"
        ],
        "title": "ProblemDetail",
        "description": "RFC 9457 problem details body returned, as ``application/problem+json``, for every web API error."
      },
...
    }
  }
}